implemented authorizations

+++ src/api/notes/handler.js
@@ -14,8 +14,9 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: creadentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({ title, body, tags, owner: creadentialId });
 
     const response = h.response({
       status: 'success',
@@ -28,8 +29,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: creadentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(creadentialId);
     return {
       status: 'success',
       data: {
@@ -40,6 +42,9 @@ class NotesHandler {
 
   async getNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -51,7 +56,11 @@ class NotesHandler {
 
   async putNoteByIdHandler(request) {
     this._validator.validateNotePayload(request.payload);
+
     const { id } = request.params;
+    const { id: creadentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, creadentialId);
     await this._service.editNoteById(id, request.payload);
 
     return {
@@ -62,6 +71,9 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: creadentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, creadentialId);
     await this._service.deleteNoteById(id);
 
     return {
